define default settings
