+++ src/components/pages/experience.tsx
@@ -10,7 +10,6 @@ import {
 
 import {
   BubbleButton,
-  // State,
   ExperienceEnv,
   SketchFunctions
 } from '../types'
