+++ src/components/pages/index.ts
@@ -2,4 +2,4 @@ export {default as About} from './about'
 export {default as Contact} from './contact'
 export {default as Projects} from './projects'
 export {default as Experience} from './experience'
-export {default as Page} from './page'
+export {default as Home} from './home'
