+++ src/components/utilities/index.ts
@@ -3,3 +3,4 @@ export * from './math'
 export * from './rendering'
 export * from './physics'
 export * from './utils'
+export * from './composites'
